agent/langgraphDemo/pyproject.toml
[project]
name = "langgraphdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.13.0",
    "langchain>=1.2.15",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.9",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
